==> 02-Associations.Rmd <==
---
title: "AACN Survey - 2022"
subtitle: "Associations"
date: "`r format(Sys.time(), '%B, %Y')`"
output:
  word_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message=FALSE, warning=FALSE)
```

```{r include=FALSE}
# Libraries ---------------------------------------------------------------

library(gtsummary)
library(haven)
library(labelled)
library(flextable)
library(glue)
library(tidyverse)

# Setup -------------------------------------------------------------------

source("00-variable_info.R", local = TRUE)
source("00-helper_functions.R", local = TRUE)

gtsummary::set_gtsummary_theme(theme_gtsummary)

flextable::set_flextable_defaults(
  table.layout = "autofit"
)

dta <- readRDS("data/raw_cleaned.rds")

labels <- labelled::var_label(dta)

# in case Respect by variables are shown independenly
new_respect_labs <- labels[vars_respect_by] %>%  map(~ paste0("Respect by:", .))

```


<br>

## Perceptions of work environment based on intent to leave

We explored the association between the Intention to leave vs the Quality of Collaboration, Communication, perceived respect and other work environment indicators such as Value, Quality of Care, Health & Safety, and Staffing levels. 

The responses in the rating items (eg. Poor, Fair, Good, Excellent) were coded to numeric [eg. 1:Poor, 2:Fair, 3:Good, to 4:Excellent] and the mean score and the standard deviation (SD) was calculated at each level of intention to leave. A higher score indicates a higher level of quality/ satisfaction/ agreement etc.



```{r eval=FALSE, include=FALSE}

### Quality of Communication and Intention to leave


### Percentage wise

new_comm_labs <- labels[vars_communication] %>%  map(~ paste0(., " rated Good or Excellent"))

by_plan <- function(dta, variable){
  
  out <- dta %>% 
    transmute(
      plan_leave = as_factor(plan_leave),
      {{variable}} := as_factor({{variable}})
    ) %>% 
    tbl_summary(by = {{variable}},
                percent = "row"
    )  
  #add_overall() 
  temp <- deparse(substitute(variable))
  out %>% 
    modify_spanning_header(
      all_stat_cols() ~ labels[[temp]]
    ) 
  
  
}


dta %>% 
  transmute(
    plan_leave = as_factor(plan_leave),
    commun_physicians = as_factor(commun_physicians)
  ) %>% 
  tbl_summary(by = commun_physicians,
              percent = "row"
  ) %>% 
  #add_overall() %>% 
  modify_spanning_header(
    all_stat_cols() ~ labels$commun_physicians
  )


dta %>% 
  filter(section == "Section C") %>% 
  transmute(
    across(all_of(vars_communication), ~ if_else(. %in% c(1,2) , 1, 0)),
    plan_leave = as_factor(plan_leave)
  ) %>% 
  set_variable_labels(.labels = labels[names(.)]) %>% 
  set_variable_labels(.labels = new_comm_labs) %>% 
  tbl_summary(
    by = plan_leave,
    statistic = all_categorical() ~ "{p}%"
    #,digits = all_categorical() ~ 1
  ) %>% 
  modify_header(label = "Communication") %>% 
  modify_spanning_header(
    all_stat_cols() ~ labels$plan_leave
  ) %>% 
  add_p() %>% 
  modify_caption(
    glue(
      "Intention to leave Vs\n
    How would you rate the quality of Communication in your unit among the following?")) %>% 
  modify_footnote(
    update = all_stat_cols() ~ glue("% Percentage of RNs who rated communication as Good or Excellent")) 

```

### Quality of Communication and Intention to leave

There is a statistically significant association between Intention to leave and quality of Communication as rated by the RNs.

Those RNs who have plans to leave within the next 12 months or 3 years, give a lower 
rating on average for the quality of collaboration between Nurses, between nurses and physicians, between Nurses and nursing managers and between nurses and the hospital administration.

```{r}
# vars_communication, vars_collaboration do a reverse coding, so higher means better 

dta %>% 
  filter(section == "Section C") %>% 
  transmute(
    across(all_of(vars_communication), ~ 5 - .),
    plan_leave = as_factor(plan_leave)
  ) %>% 
  set_variable_labels(.labels = labels[names(.)]) %>% 
  tbl_summary(
    by = plan_leave,
    type = everything() ~ "continuous"
  ) %>% 
  modify_header(label = "Communication") %>% 
  modify_spanning_header(
    all_stat_cols() ~ labels$plan_leave
  ) %>% 
  add_p() %>% 
  modify_caption(
    glue(
      "Intention to leave Vs\n
    How would you rate the quality of Communication in your unit among the following?")) %>% 
  modify_footnote(
    update = all_stat_cols() ~ glue("Mean (SD) / Mean of scores ranging from 1:Poor, 2:Fair, 3:Good, to 4:Excellent; a higher score indicates a higher level of quality of Communication")) 


```


### Quality of Collaboration and Intention to leave

There is a statistically significant association between Intention to leave and quality of collaboration as rated by the RNs.

Those RNs who have plans to leave within the next 12 months or 3 years, give a lower 
rating on average for the quality of collaboration between Nurses, between nurses and physician, between Nurses and nursing managers and between nurses and the hospital administration.


```{r}
# vars_communication, vars collaboration do a reverse coding, so higher means better 

dta %>% 
  filter(section == "Section C") %>% 
  transmute(
    across(all_of(vars_collaboration), ~ 5 - .),
    plan_leave = as_factor(plan_leave)
  ) %>% 
  set_variable_labels(.labels = labels[names(.)]) %>% 
  tbl_summary(
    by = plan_leave,
    type = everything() ~ "continuous"
  ) %>% 
  modify_header(label = 'Collaboration') %>% 
  modify_spanning_header(
    all_stat_cols() ~ labels$plan_leave
  ) %>% 
  add_p() %>% 
  modify_footnote(
    update = all_stat_cols() ~ glue::glue("Mean (SD) / Mean of scores ranging from 1:Poor, 2:Fair, 3:Good, to 4:Excellent; a higher score indicates a higher level of quality of Collaboration")
  ) %>% 
  modify_caption(
    glue(
      "Intention to leave Vs\n
    How would you rate the quality of Collaboration in your unit among the following?")
  )
```


### Respect and Intention to leave

There is a statistically significant association between Intention to leave and the perceived respect for nurses.

Those RNs who have plans to leave within the next 12 months or 3 years, give a lower 
rating on average for the respect by Other nurses, Physicians, Other Health Care colleagues, Unit managers and Hospital administration.    


```{r}

dta %>% 
  filter(section == "Section C") %>% 
  transmute(
    across(all_of(vars_respect_by), ~ 5 - .),
    plan_leave = as_factor(plan_leave)
  ) %>% 
  set_variable_labels(.labels = labels[names(.)]) %>% 
  tbl_summary(
    by = plan_leave,
    type = everything() ~ "continuous"
  ) %>% 
  modify_header(label = "**Respect by**") %>% 
  modify_spanning_header(
    all_stat_cols() ~ labels$plan_leave
  ) %>% 
  add_p() %>% 
  modify_footnote(
    update = all_stat_cols() ~ glue::glue("Mean (SD) / Mean of scores ranging from 1:Poor, 2:Fair, 3:Good, to 4:Excellent; a higher score indicates a higher level of respect")
  ) %>% 
  modify_caption(
    glue(
      "Intention to leave Vs\n
    In your unit how would you rate the respect for nurses by each of the following?"))


```


### Intention to leave Vs perceptions of work environments [Value, Quality of Care, Health & Safety, Staffing]


Those RNs who have plans to leave within the next 12 months or 3 years, give a lower 
rating on average when describing the quality of care for patients in their unit, compared to the RNs that have no plans to leave within the next 3 years.

They also tend to agree less (on a scale of 1 (Strongly Disagree) to 4 (Strongly Agree) that they are "recognised for the value they bring to their work environment", and that their "Organisation values their health and safety". Lastly, they indicated that they experience more frequently n average, moral distress compared to the RNs that have no plans to leave within the next 3 years.


```{r}

dta %>% 
  filter(section == "Section C") %>% 
  transmute(
    plan_leave = as_factor(plan_leave), 
    across(
      c(quality_care, hwes_12_unit, value_health_safety, moral_distress), ~ 5 - .
    ),
    staffing_right = if_else(staffing_right >= 4, 1, 0)
  ) %>% 
  set_variable_labels(.labels = labels[names(.)]) %>% 
  tbl_summary(
    by = plan_leave,
    type = list( c(quality_care, hwes_12_unit, value_health_safety,moral_distress) ~ "continuous"),
    missing = "no"
  ) %>% 
  modify_header(label = "**Work Environment**") %>% 
  modify_spanning_header(
    all_stat_cols() ~ labels$plan_leave
  ) %>% 
  modify_caption(
    glue(
      "Intention to leave Vs Perceptions of work environments")) %>% 
  add_p() %>% 
  modify_footnote(
    update = all_stat_cols() ~ glue::glue("n(%) for Right Staffing All the time, or more than 75% of the time.\n
  Mean(SD) for the rest. Mean of scores ranging from   
  1:Poor, 2:Fair, 3:Good, to 4:Excellent for 'Quality of Care';  
  1:Strongly disagree, 2:Disagree, 3:Agree, to 4:Strongly agree for 'RNs are recognized for the value' and 'My organization values my health and safety';  
  1:Very Rarely, 2:Occasionally, 3:Frequently, to 4:Very Frequently for 'Moral Distress'")) 


```



## Relationships with __AACN Standards__

The tables below present the Spearman correlation coefficient (r) for the linear associations between the six AACN standards, and the Wοrk Environment indicators: __Job satisfaction__ , __Collaboration__, __Communication__, __Intention to leave__ and __Moral Distress__.


### AACN Standards with respect to the Organisation


There are moderate and positive correlations between the AACN standards and the Communication environment (correlation ranging from 0.24 ~ 0.34 ), where "Communication between Nurses and Hospital Administration" having the highest correlation with the AACN standards (correlation r ranging from 0.42 ~ 0.54), indicating that higher communication is associated with increased AACN standards across all areas, not only for Skilled communication and True collaboration but also for Effective Decision making (r = 0.48), Appropriate staffing (r = 0.45), Meaningful recognition (r = 0.48) and Authentic Leadership (r = 0.54).

A very similar level of positive correlation is shown for the Collaboration in the work environment and the six AACN standards.

With respect to the moral distress level, there is a moderate and negative association with the six AACN standards, where lower moral distress is associated with higher level of Skilled Communication (r= -0.22), True Collaboration (r=-0.22), Effective Decision-Making (r = -0.19), Appropriate Staffing (r = -0.14), Meaningful Recognition (r = -0.2) and Authentic Leadership (r = -0.23).

Moreover, satisfaction with being a Registered Nurse and satisfaction with the current job, have a low and positive association  with the six AACN standards.

The level of respect they feel is moderately and positively associated with the six AACN standards, indicating that the higher the level of respect the RNs feel, is associated with higher AACN standards across all areas of the AACN.

<br>

```{r}

dta %>% 
  mutate(
    # make inappropriate staffing = 1 so the same direction as with 
    # AACN scales (negative direction)
    staffing_right = ifelse(staffing_right >= 4, 0, 1)
  ) %>% 
  set_variable_labels(
    staffing_right = "Appropriate staffing in the unit(more than 75% of time or all of the time)"
  ) %>% 
  set_variable_labels(.labels = new_respect_labs) %>% 
  tbl_cross_correlation(
    .colsx = 
      c( 
        all_of(vars_communication),
        all_of(vars_collaboration),
        moral_distress,
        quality_care,
        # staffing_right,
        
        satisfied_nurse,
        satisfied_overall,
        all_of(vars_respect_by)
      ),
    .colsy = all_of(dim__scales_org)
  ) %>% 
  rename("Work Environment" = colsx) %>% 
  flextable() %>% 
  flextable::add_header_row( 
    values = c("", "AACN Standards (Organisation level)"), 
    colwidths = c(1,6)
  ) %>% 
  add_footer_row(
    values = c("Note: Spearman correlation coefficients (r). All correlations are significant at the 0.01 level (2-tailed). ", ""), 
    colwidths = c(6, 1) 
  ) %>% 
  add_footer_row(
    values = c("AACN, American Association of Critical-Care Nurses;"), 
    colwidths = 7
  ) %>% 
  set_caption(
    "Correlation (Spearman r) between AACN Standards (Organisation level), 
    and Job satisfactin , Collaboration, Communication, Intention to leave 
    and Moral Distress")

```


### AACN Standards with respect to the Unit

A similar association is observed as in the AACN standards in the Organisation.    

<br>

```{r}

dta %>% 
  mutate(
    # make inappropriate staffing = 1 so the same direction as with AACN scales (negative direction)
    staffing_right = ifelse(staffing_right >= 4, 0, 1)
  ) %>% 
  set_variable_labels(
    staffing_right = "Appropriate staffing in the unit(more than 75% of time or all of the time)"
  ) %>% 
  set_variable_labels(.labels = new_respect_labs) %>% 
  tbl_cross_correlation(
    .colsx = 
      c( 
        all_of(vars_communication),
        all_of(vars_collaboration),
        moral_distress,
        quality_care,
        #staffing_right,
        
        satisfied_nurse,
        satisfied_overall,
        all_of(vars_respect_by)
      ),
    .colsy = all_of(dim__scales_unit)
  ) %>% 
  rename(" " = colsx) %>% 
  flextable() %>% 
  flextable::add_header_row(
    values = c("", "AACN Standards (Unit level)"), 
    colwidths = c(1,6)
  ) %>% 
  add_footer_row(
    values = c("Note: Spearman correlation coefficients (r). 
               All correlations are significant at the 0.01 level (2-tailed)."),
    colwidths = c(7) 
  ) %>% 
  add_footer_row(
    values = c("AACN, American Association of Critical-Care Nurses;"), 
    colwidths = 7
  ) %>% 
  set_caption(
    "Correlation (Spearman r) between AACN Standards (Unit level), 
    and Job satisfactin , Collaboration, Communication, Intention to leave 
    and Moral Distress")



```


## Association of Perceived Respect and HWE implementation level

We explored the association between the HWE implementation level, vs the rating of the respect level the RNs perceive.  

The rating items (eg. Poor, Fair, Good, Excellent) were coded to numeric [eg. 1:Poor, 2:Fair, 3:Good, to 4:Excellent] and the mean score and the standard deviation (SD) was calculated at each level of intention to leave. A higher score indicates a higher level of quality/ satisfaction/ agreement etc..

<br>

Results:    

<br>

There is a statistically significant association between the perceive level of respect for RNs and whether their organisation has implemented Healty Work Environment standards. (p<0.001).

RNs who stated that their organisation has Fully Implemented some Healthy Work Environment standards, rate higher on average, the respect by Other nurses, by Physicians, by Other health care colleagues, by unit nursing managers, and by the hospital administration, compared to the the RNs who stated that their organisation has not all implemented any HWE standards.

<br>

```{r}

dta %>% 
  mutate(
    across(all_of(vars_respect_by), ~ 5 - .)
  ) %>% 
  set_variable_labels(.labels = labels[vars_respect_by]) %>% 
  select(
    implement_hwes_unit,
    all_of(vars_respect_by)
  ) %>% 
  mutate(implement_hwes_unit = as_factor(implement_hwes_unit)) %>% 
  tbl_summary(
    by = implement_hwes_unit,
    missing = "no",
    type = everything()~ "continuous"
  ) %>% 
  add_p() %>% 
  modify_header(label = "-Respect by-") %>% 
  modify_caption(
    "Correlations between implementation of AACN HWE standards and 
    Respect satisfaction levels") %>% 
  modify_spanning_header(
    all_stat_cols() ~ labels$implement_hwes_org
  ) %>% 
  modify_footnote(
    p.value ~ "One-way ANOVA\n
  Note: Mean of scores (SD) ranging from 1:Poor, 2:Fair, 3:Good, to 4:Excellent" 
  )

```



## Association between __Perceived Respect__ and work environment indicators: __Job satisfaction__, __Communication__, __Collaboration__, __Moral Distress__ and __Intention to leave__.    

<br>

There are moderate to strong and positive correlations between __Communication__ satisfaction and the perceived respect items, indicating that increased communication is associated with increased perceived respect for nurses. 

Similarly, there are moderate to strong and positive correlations between __Collaboration__ satisfaction and the perceived respect items, indicating that increased satisfaction in the collaboration area, is associated with increased perceived respect for nurses. 

Intention to leave within 1 or 3 years, is negatively (low: r: -0.12 ~ -0.22) associated with perceived respect for nurses. Nurses who indicated that they have intention to leave within 1 or 3 years, tend to have lower satisfaction with the respect they perceive.

Moral distress, is negatively (low: r: -0.16 ~ -0.25) associated with perceived respect for nurses. The higher the moral distress experienced by RNs, the lower satisfaction with the respect they perceive.

Satisfaction with the current job is moderately and positively associated with increased perceived respect for RNs (r: 0.31 ~ 0.37)


```{r}

dta %>% 
  mutate(
    # make inappropriate staffing = 1 so the same direction as 
    # with AACN scales (negative direction)
    plan_leave = ifelse(plan_leave == 3, 1, 0)
  ) %>% 
  set_variable_labels(
    plan_leave = "Intention to leave within 1 or 3 years"
  ) %>% 
  tbl_cross_correlation(
    c( 
      all_of(vars_communication),
      all_of(vars_collaboration),
      plan_leave,
      moral_distress,
      # job satisfaction
      satisfied_nurse,
      satisfied_overall
    ),
    all_of(vars_respect_by)
  ) %>% 
  rename(" Work Environment " = colsx) %>% 
  flextable() %>% 
  flextable::add_header_row(
    values = c("", 
               "In your unit how would you rate the respect for nurses 
               by each of the following?"), 
    colwidths = c(1,5)
  ) %>% 
  add_footer_row(
    values = c("Note: Spearman correlation coefficient (r). 
               All correlations are significant at the 0.01 level (2-tailed). "), 
    colwidths = c(6) 
  ) %>% 
  set_caption(
    "Correlation between Respect and Job satisfaction, Communication, Collaboration, Moral DIstress and Intention to leave")

```


###  -	C17 with c18

# C17. On the whole, how satisfied are you with your current job?
# C18. Do you plan to leave your present nursing position?
# C18a. How likely would each of the following be to influence you to reconsider your plans to leave your present position? 
# this one is many questions

D1. Gender 
D4. Years of total nursing experience 
D5. Years of nursing experience in critical/intensive care unit
D6. Current position
D7. Place of current work:    
D8. Type of ICU
D9. Country of residence: 


<br>



## Association of Demographics Vs C17. On the whole, how satisfied are you with your current job?


```{r}

dta %>%
  filter(section == "Section C") %>%
  select(satisfied_overall, 
         # demo
         gender, position, experience, 
         experience_icu,  icu_type, country
  ) %>%
  mutate(
    across(where(is.labelled), as_factor)
  ) %>% 
  set_variable_labels(.labels = labels[names(.)]
                      #plan_leave = "Intention to leave within 1 or 3 years"
  ) %>%
  tbl_summary(
    by = satisfied_overall,
    missing = "no",
    percent = "row"
  ) %>%
  add_p(
    test = list(all_categorical() ~ "chisq.test")
  ) %>% 
  #modify_header(label = "") %>% 
  modify_caption(
    "Associations between Demographics and Intention to Leave") %>% 
  modify_spanning_header(
    all_stat_cols() ~ labels$satisfied_overall
  ) 

```

<br>


## Association of Demographics Vs C18. Do you plan to leave your present nursing position?

<br>


```{r}

dta %>%
  filter( section == "Section C" ) %>%
  select( plan_leave, gender, position, experience, experience_icu,  icu_type, country ) %>%
  mutate( across(where(is.labelled), as_factor) ) %>% 
  set_variable_labels( .labels = labels[names(.)] ) %>%
  tbl_summary( by = plan_leave, missing = "no", percent = "row" ) %>%
  add_p( test = list(all_categorical() ~ "chisq.test") ) %>% 
  #modify_header(label = "") %>% 
  modify_caption( "Associations between Demographics and Intention to Leave" ) %>% 
  modify_spanning_header( all_stat_cols() ~ labels$plan_leave ) 

```


## Association of Demographics Vs C18a. How likely would each of the following be to influence you to reconsider your plans to leave your present position? 

Responses to eleven (11) questions, were in the range of __Very likely, Somewhat likely, Not very likely__ and __Not likely at all__. The associations were performed on the event of response of __Very Likely__ to be influenced by the said factor.


```{r include=FALSE}
by_demo1 <- function(dta, demo_var){
  
  out <- dta %>% 
    filter(section == "Section C") %>%
    select( starts_with("influence_"), {{demo_var}} ) %>% 
    mutate( across(starts_with("influence_"), ~if_else(. == 1, 1, 0)) ) %>% 
    mutate(across(where(is.labelled), as_factor)) %>% 
    set_variable_labels(.labels = labels[names(.)]) %>% 
    tbl_summary( by = {{demo_var}}, missing = "no" ) %>%
    add_p( test = list(all_categorical() ~ "chisq.test") ) %>% 
    modify_header(label = "Very likely to be influenced by:") %>% 
    modify_caption( "Associations between Demographics and 'How likely would each of the following be to influence you to reconsider your plans to leave your present position?'") 
  
  temp <- deparse(substitute(demo_var))
  out %>% 
    modify_spanning_header( all_stat_cols() ~ labels[[temp]] ) 
  
}


```

<br>

## Gender 

<br>

```{r}
by_demo1(dta, gender)

```
<br>

## Current position 

<br>

```{r}
by_demo1(dta, position)

```
<br>

## Place of current work 

<br>

```{r}
by_demo1(dta, workplace)

```
<br>

## Type of ICU 

<br>

```{r}
by_demo1(dta, icu_type)

```
<br>

## Country 

<br>

```{r}
by_demo1(dta, country)

```

<br>

```{r}
t1 <- dta %>% 
  filter(section == "Section C") %>%
  tbl_cross_correlation(starts_with("influence_"), experience)

t2 <- dta %>% 
  filter(section == "Section C") %>%
  tbl_cross_correlation(starts_with("influence_"), experience_icu)

left_join(t1, t2) %>% 
  rename("Likelihood of influenced by:" = colsx) %>% 
  flextable() %>% 
  flextable::add_header_row(
    values = c("", 
               "Correlation with"), 
    colwidths = c(1,2)
  ) %>% 
  add_footer_row(
    values = c("Note: Spearman correlation coefficient (r). 
               * p < 0.05, ** p < 0.01 "), 
    colwidths = c(3) 
  ) %>% 
  set_caption(
    "Correlation between  `How likely would each of the following be to influence you to reconsider your plans to leave your present position?` and `Years of Experience (Nursing / ICU)`")

```



## Associations A1, A3, A5, A 6 with C17. On the whole, how satisfied are you with your current job?


<br>


```{r}
dta %>% 
  filter(section == "Section C") %>% 
  select(
    implement_hwes_unit, implement_hwes_org, satisfied_nurse, quality_care,
    satisfied_overall
  ) %>% 
  mutate(across(where(is.labelled), as_factor),
         quality_care = fct_drop(quality_care)) %>% 
  tbl_summary( by = satisfied_overall,
               missing = "no",
               percent = "row"
  ) %>% 
  add_overall() %>% 
  add_p( test = list(all_categorical() ~ "chisq.test") ) %>% 
  modify_header(all_stat_cols() ~ "**{level}**",label = "**Healthy Work Environmnet**") %>% 
  modify_caption( "Associations of 'How satisfied are you with your current job' and ' Implementation of HWE standards' " ) %>% 
  modify_spanning_header( all_stat_cols() ~ labels$satisfied_overall ) 

# Correlations
# dta %>% 
#   filter(section == "Section C") %>%
#   tbl_cross_correlation(c(satisfied_nurse, quality_care), satisfied_overall) %>% 
#   rename(" " = colsx) %>% 
#   flextable()

```


## Associations of A1, A3, A5, A 6 with C18. Do you plan to leave your present nursing position?


<br>


```{r}
dta %>% 
  filter(section == "Section C") %>% 
  select(
    implement_hwes_unit, implement_hwes_org, satisfied_nurse, quality_care,
    plan_leave
  ) %>% 
  mutate(across(where(is.labelled), as_factor),
         quality_care = fct_drop(quality_care)
         ) %>% 
  tbl_summary( by = plan_leave,
               missing = "no",
               percent = "row"
  ) %>% 
  add_overall() %>% 
  add_p( test = list(all_categorical() ~ "chisq.test") ) %>% 
  modify_header(all_stat_cols() ~ "**{level}**",label = "**Healthy Work Environmnet**") %>% 
    modify_caption( "Associations of 'Intention to leave' and ' Implementation of HWE standards' " ) %>% 
  modify_spanning_header( all_stat_cols() ~ labels$plan_leave ) 

```


<br>
