==> 03-Reliability.Rmd <==
---
title: "Reliability analysis - AACN survey"
author: "Lefkios Paikousis"
date: "`r format(Sys.time(), '%B, %Y')`"
output:
  word_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
```

```{r include=FALSE}
# Libraries ---------------------------------------------------------------

library(gtsummary)
library(haven)
library(labelled)
library(flextable)
library(glue)
library(tidyverse)

# Setup -------------------------------------------------------------------

source("00-variable_info.R", local = TRUE)
source("00-helper_functions.R", local = TRUE)

gtsummary::set_gtsummary_theme(theme_gtsummary)

flextable::set_flextable_defaults(
  table.layout = "autofit"
)

dta <- readRDS("data/dta_retest.rds")

var_label(dta) %>% keep(is.null)

labels <- labelled::var_label(dta)

# in case Respect by variables are shown independently
new_respect_labs <- labels[vars_respect_by] %>%  map(~ paste0("Respect by:", .))

new_vars_hwes_labs <- labels[vars_hwes_org] 
names(new_vars_hwes_labs) <- str_remove(names(new_vars_hwes_labs), "_org")

```



```{r process, include=FALSE}

dta %>% count(id_comb, id, test,sort= T)

dta_split <- 
  dta %>% 
  arrange(id_comb,test) %>% 
  #select(id_comb,test, all_of(vars_collaboration)) %>% 
  split(.$test)


test_retest <- function(vars_vector){
  
  cor(
    x = dta_split$test[vars_vector],
    y = dta_split$retest[vars_vector], 
    use = "pairwise.complete.obs",
    method = "spearman") %>% 
    diag() %>% 
    enframe() %>% 
    transmute(
      var = name,
      r  = round(value, 2)
    )
  
}


icc_var <- function(var){
  
obj <- 
  map_dfr(dta_split, ~select(., !!!var) %>% pull) %>% 
  as.matrix() %>% 
  psych::ICC(lmer = TRUE)

obj$results[3,]$ICC %>% setNames(var)

}

icc_var("collab_nurses")


test_retest <- function(vars){
  
  vars %>% 
  set_names() %>% 
  map_dbl(icc_var) %>% 
  tibble::enframe("var", "icc") %>% 
  mutate(icc = round(icc,2))
  
}

# check
test_retest(vars_communication)


```


## Test - Retest reliability

We estimated the test-retest reliability using the Intra-Class Correlation coefficient (ICC).  
The type of ICC used is the __Two-Way mixed effects, consistency, single rater/measurement__ as defined by _[Shrout & Fleiss (1979)]_ convention. ICC values less than 0.5, between 0.5 and 0.75, between 0.75 and 0.9, and greater than 0.90 are indicative of poor, moderate, good, and excellent reliability, respectively __(Koo TK, Li MY.2016)__.   


The ICC is estimated for the items in Section A, Section B (HWE scale) and for a range of items in Section C namely the __Collaboration/ Communication__ items, the __perceived respect__ for nurse, and the ratings of __manager skills__.  


Ref:: [Koo TK, Li MY. A Guideline of Selecting and Reporting Intraclass Correlation Coefficients for Reliability Research [published correction appears in J Chiropr Med. 2017 Dec;16(4):346]. J Chiropr Med. 2016;15(2):155-163. doi:10.1016/j.jcm.2016.02.012]


### Section A


There is moderate to good reliability (ICC: 0.76 ~ 0.82) in the items of Section A.    

<br>

```{r}

vars_sectionA <- c(
  c("satisfied_nurse", "advice_nurse", "quality_care")
  
  #"aware_hwes", "implement_hwes_unit", "implement_hwes_org")
)


test_retest(vars_sectionA) %>% 
  mutate(var = recode(var, !!!unlist(labels))) %>% 
  rename("How would you rate the quality of communication in your unit among the following?" = var) %>% 
  flextable() %>% 
  set_caption("Test-retest reliability in the Communication items")


```


### Section B (HWE)

The test retest reliability of the HWE items was tested in all 16 items - for both _in your Organisation_ and _In your unit_.


There is moderate to good reliability (ICC: 0.71 ~ 0.82) in the items of the HWE scale.    
<br>


```{r echo=FALSE}
org <- test_retest(vars_hwes_org) 
unit <- test_retest(vars_hwes_unit)

bind_rows(org, unit) %>% 
  extract(var, c("var", "type"),"(.+\\d)_(org|unit)") %>% 
  mutate(var = fct_inorder(var)) %>% 
  spread(type, icc) %>% 
  mutate(var = recode(var, !!!unlist(new_vars_hwes_labs))) %>% 
  rename("HWE item" = var, "Organisation" = org, "Unit" = unit) %>% 
  flextable() %>% 
  add_header_row(values = c("", "In your"), colwidths = c(1, 2)) %>% 
  add_footer_lines("Table shows ICC coefficient") %>% 
  set_caption("Test-retest reliability in the HWE items")

```

### Section C

- Communication

There is moderate reliability (ICC: 0.71 ~ 0.78) in the Communication scale.  

<br>

```{r }
# bind_rows(
#   Collaboaration = test_retest(vars_collaboration),
#   Communication = test_retest(vars_communication),
#   .id = "type"
# ) %>% 
test_retest(vars_communication) %>% 
  mutate(var = recode(var, !!!unlist(labels))) %>% 
  rename("How would you rate the quality of communication in your unit among the following?" = var) %>% 
  flextable() %>% 
  merge_v(j = 1) %>% 
  set_caption("Test-retest reliability in the Communication items")

```


<br>

- Collaboration

There is moderate to good reliability (ICC: 0.71 ~ 0.82) in the Collaboration items.   


```{r }
test_retest(vars_collaboration) %>% 
  mutate(var = recode(var, !!!unlist(labels))) %>% 
  rename("How would you rate the quality of collaboration in your unit among the following?" = var) %>% 
  flextable() %>% 
  set_caption("Test-retest reliability in the Collaboration items")


```

<br>

- Respect by

There is moderate reliability (ICC: 0.72 ~ 0.78) in the perceived respect items.   


```{r }
test_retest(vars_respect_by) %>% 
  mutate(var = recode(var, !!!unlist(labels))) %>% 
  rename("In your unit how would you rate the respect for nurses by each of the following?" = var) %>% 
  flextable() %>% 
  set_caption("Test-retest reliability in the perceived respect items")

```


<br>

- Rate Manager skills

There is moderate (ICC: 0.70 ~ 0.76) in the Manager's skills items.   


```{r }
test_retest(vars_skills_manager) %>% 
  mutate(var = recode(var, !!!unlist(labels))) %>% 
  rename("Please rate the skill of your unit Nursing managers in the following areas" = var) %>% 
  flextable() %>% 
  set_caption("Test-retest reliability for the Manager SKills rating items")

```

<br>

- Rate Hospital Administration skills

There is moderate (ICC: 0.66 ~ 0.77) in the Hospital Administration's skills items.   

```{r }
test_retest(vars_skills_admin) %>% 
  mutate(var = recode(var, !!!unlist(labels))) %>% 
  rename("Please rate the skill of your hospital nursing administrators in the following" = var) %>% 
  flextable() %>% 
  set_caption("Test-retest reliability for the hospital nursing administrators' Skills rating items")

```


## Internal Consistency

Internal consistency was evaluated using the Cronbach's alpha index. The alpha index was calculated for the rating items in Section B (HWE scale) and the Section C's items for the __Collaboration/ Communication__ items, the __perceived respect__ for nurse, and the ratings of __manager skills__.   

There is good to excellent internal consistency in all groups of items (alpha range: 0.78 ~ 0.97)

<br>


```{r include=FALSE}

cronbach <- function(.data, .vars, .check.keys = FALSE){
  
  
  stopifnot(inherits(.data, "data.frame"))
  #stopifnot(length(.digits)==1 | is.integer(.digits))
  
  exprX <- rlang::enquo(.vars)
  
  posX <- tidyselect::eval_select(exprX, data = .data)
  
  
  if (length(posX) == 0) {
    rlang::abort(glue::glue("`.vars` must select at least one column."))
  }
  
  # this also renames of you pas smthng like .vars = c(New_name = old_name)
  dta <- rlang::set_names(.data[posX], names(posX))
  
  n_items <- length(posX)
  
  alpha <- dta %>% 
    psych::alpha(check.keys = .check.keys) %>% .$total %>% .$std.alpha
  
  tibble(
    n = n_items,
    alpha = alpha
    
  )
}
 
dta <- readRDS("data/raw_cleaned.rds")

missing_removed <- readxl::read_xlsx("AACN_survey-missing_values.xlsx", sheet = "MissingRemoved")

dta <- dta %>% 
  filter(id %in% missing_removed$Responseid)

cronbach(dta, all_of(vars_sectionA))

```


```{r}

list_of_alphas <- 
  list(
    org   = cronbach(dta, vars_hwes_org),
    unit  = cronbach(dta, vars_hwes_unit),
    comm  = cronbach(dta, vars_communication),
    coll  = cronbach(dta, vars_collaboration),
    resp  = cronbach(dta, vars_respect_by),
    mskil = cronbach(dta, vars_skills_manager),
    askil = cronbach(dta, vars_skills_admin)
    
  ) 


list_of_alphas %>% 
  bind_rows(.id = "group") %>% 
  mutate(alpha = round(alpha, 2)) %>% 
  mutate(
    group = fct_recode(group, 
                       "Healthy Work Environmnet - Organisation" = 'org',
                       "Healthy Work Environmnet - Unit" = 'unit',
                       "Quality of communication in your unit" = 'comm',
                       "Quality of collaboration in your unit" = 'coll',
                       "In your unit how would you rate the respect for nurses " = 'resp',
                       "Skills of your unit Nursing managers" = 'mskil',
                       "Skills of your hospital nursing administrator" = 'askil'
                       
    )
  ) %>% 
  rename(
    "Group of items" = group,
    "N (Items)" = n,
    "Cronbach's alpha" = alpha
  ) %>% 
  flextable() %>% 
  set_caption("Internal Consistency (Cronbach's alpha) for the items in the AACN survey")

```
